Batch 3 slice 1: specification v0.46, streams and TCP under T4
One amendment, landed as one change: the active specification is v0.46
over v0.45's archived bytes, the approval record is appended, the
identity module regenerated, the qualification review moved. Output is
OutputStream; FileFactory, FilePermit, command.files and reserve_file
are HandleFactory, HandlePermit, command.handles and reserve_handle,
because sockets draw from the same descriptor table. command.stdin at
ordinal 5 supplies an InputStream and read_next reads it through the
one submit-then-join lowering: the ring's IORING_OP_READ at offset -1
on Linux, the shared adapter elsewhere, ReadFile on Windows. TCP enters
the language whole: SocketAddress with its two pure constructors,
TcpListener, TcpConnection as a system-declared struct of TcpReceive and
TcpSend, the three outcome enums handing the permit back, tcp_listen,
tcp_accept, tcp_connect, receive_next, send_once, close_connection and
close_listener under [SYS-15] to [SYS-18]; their runtime routes are
slice 2 and a module that submits one stops at target qualification
with a missing mapping, never a source rejection. Eleven conformance
cases are added and thirty-one respelled with their verdicts unchanged.

Two defects found on the way: the lowering decoded a system constructor
ordinal against the active inventory rather than the unit's own, which
a second nominal-record block exposed; and no checked borrow form
carried a field path for a non-buffer type, which &uniq
connection.receive needed.

+++ compiler/src/resolution/mod.rs
@@ -141,17 +141,20 @@ impl PreludeDeclarationId {
 }
 
 /// Dense identity of one normative [SYS-2] system declaration record.
+///
+/// The ordinal is a `u16` because v0.46's inventory is three hundred and seven
+/// records; it was a `u8` while the inventory fitted one byte.
 #[derive(Clone, Copy, Debug, Eq, Hash, Ord, PartialEq, PartialOrd)]
-pub struct SystemDeclarationId(u8);
+pub struct SystemDeclarationId(u16);
 
 impl SystemDeclarationId {
-    pub(crate) const fn new(ordinal: u8) -> Self {
+    pub(crate) const fn new(ordinal: u16) -> Self {
         Self(ordinal)
     }
 
     /// Returns the zero-based `system_declaration_ordinal` in [SYS-2] preorder.
     #[must_use]
-    pub const fn ordinal(self) -> u8 {
+    pub const fn ordinal(self) -> u16 {
         self.0
     }
 }
